fix(store): address 18 review issues from PR #29
Critical (4):
- C1: Replace format! SQL injection in cleanup_stale_conn with temp table + parameter binding
- C2: Replace open_db().unwrap() panic in serve.rs with match-based graceful degradation
- C3: Improve u64_to_i64 documentation clarifying counter-only usage and saturation safety
- C4: Add schema versioning framework with init_schema skip for existing DBs and migration path

Warning (8):
- W1: Add BEGIN IMMEDIATE transaction to save_promotion_counters_conn
- W2: Remove result IS NULL from success count in observation stats queries
- W3: Document DELETE+INSERT score_history pattern within transaction
- W4: Replace INSERT OR REPLACE with ON CONFLICT preserving first_seen in skill_attribution
- W5: Extract pad_date() helper to eliminate duplicated date padding logic
- W6: Add error logging to ingest_to_memory on DB open/transaction failures
- W7: Replace filter_map(|r| r.ok()) with explicit error logging in metrics load
- W8: Add state_json non-object guard in list_all_pipelines_conn

Info (6):
- I1: Remove unused _is_new variable from open_harness_db
- I2: Skip PRAGMA/DDL for existing DBs via schema version check (covered by C4)
- I3: Replace i32::MAX with i64::MAX for query_all_records_conn unlimited query
- I4: Add idx_orch_agents_run index for run_id lookups (covered by C4)
- I5: eprintln retained (project does not use tracing)
- I6: Update effective_sources tests to include mem in expected values

+++ src/store/evolution.rs
@@ -82,7 +82,7 @@ pub fn query_recent_records_conn(
 
 /// Query all evolution records.
 pub fn query_all_records_conn(conn: &Connection) -> io::Result<Vec<EvolutionRecord>> {
-    query_recent_records_conn(conn, i32::MAX as usize)
+    query_recent_records_conn(conn, i64::MAX as usize)
 }
 
 #[cfg(test)]
